# swagger_docs.json
{"openapi":"3.1.0","info":{"title":"AI DSS Sustainable Farming API","description":"\n    AI-Powered Decision Support System\n    for Sustainable Farming.\n\n    Features:\n    - Crop Recommendation\n    - QSSM Sustainability Score\n    - Environmental Intelligence\n    - Voice-based Advisory\n    - Government Scheme Matching\n    ","version":"1.0.0"},"paths":{"/compute-qssm":{"post":{"summary":"Compute","operationId":"compute_compute_qssm_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/knowledge/search/practices":{"get":{"tags":["Knowledge"],"summary":"Search Practices","operationId":"search_practices_knowledge_search_practices_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/knowledge/search/schemes":{"get":{"tags":["Knowledge"],"summary":"Search Schemes","operationId":"search_schemes_knowledge_search_schemes_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/knowledge/search/soil":{"get":{"tags":["Knowledge"],"summary":"Search Soil","operationId":"search_soil_knowledge_search_soil_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/knowledge/search/pest":{"get":{"tags":["Knowledge"],"summary":"Search Pest","operationId":"search_pest_knowledge_search_pest_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/farmer":{"post":{"tags":["Farmer Management"],"summary":"Create Farmer","operationId":"create_farmer_farmer_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmerCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/farm":{"post":{"tags":["Farm Management"],"summary":"Create Farm","operationId":"create_farm_farm_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/environment":{"get":{"tags":["Environment Data"],"summary":"Get Environment","operationId":"get_environment_environment_get","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","title":"Lat"}},{"name":"lon","in":"query","required":true,"schema":{"type":"number","title":"Lon"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recommendation":{"post":{"tags":["Recommendations"],"summary":"Get Recommendation","operationId":"get_recommendation_recommendation_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/history":{"post":{"tags":["Farm History"],"summary":"Create History","operationId":"create_history_history_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HistoryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/history/{farm_id}":{"get":{"tags":["Farm History"],"summary":"Get History","operationId":"get_history_history__farm_id__get","parameters":[{"name":"farm_id","in":"path","required":true,"schema":{"type":"integer","title":"Farm Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/register":{"post":{"tags":["Authentication"],"summary":"Register User","operationId":"register_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"tags":["Authentication"],"summary":"Login User","operationId":"login_user_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/voice":{"post":{"tags":["Voice Assistant"],"summary":"Voice Endpoint","operationId":"voice_endpoint_voice_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoiceRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FarmCreate":{"properties":{"farmer_id":{"type":"integer","title":"Farmer Id"},"soil_type":{"type":"string","title":"Soil Type"},"soil_ph":{"type":"number","title":"Soil Ph"},"organic_carbon":{"type":"number","title":"Organic Carbon"},"farm_size":{"type":"number","title":"Farm Size"},"irrigation_type":{"type":"string","title":"Irrigation Type"},"water_source":{"type":"string","title":"Water Source"},"gps_lat":{"type":"number","title":"Gps Lat"},"gps_lon":{"type":"number","title":"Gps Lon"}},"type":"object","required":["farmer_id","soil_type","soil_ph","organic_carbon","farm_size","irrigation_type","water_source","gps_lat","gps_lon"],"title":"FarmCreate"},"FarmerCreate":{"properties":{"name":{"type":"string","title":"Name"},"phone":{"type":"string","title":"Phone"},"district":{"type":"string","title":"District"},"village":{"type":"string","title":"Village"},"language":{"type":"string","title":"Language"}},"type":"object","required":["name","phone","district","village","language"],"title":"FarmerCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HistoryCreate":{"properties":{"farm_id":{"type":"integer","title":"Farm Id"},"crop":{"type":"string","title":"Crop"},"yield_value":{"type":"number","title":"Yield Value"},"fertilizer":{"type":"number","title":"Fertilizer"},"pesticide":{"type":"number","title":"Pesticide"},"profit":{"type":"number","title":"Profit"}},"type":"object","required":["farm_id","crop","yield_value","fertilizer","pesticide","profit"],"title":"HistoryCreate"},"RecommendationResponse":{"properties":{"recommended_crop":{"type":"string","title":"Recommended Crop"},"sustainability_score":{"type":"number","title":"Sustainability Score"}},"type":"object","required":["recommended_crop","sustainability_score"],"title":"RecommendationResponse"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserLogin":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserLogin"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VoiceRequest":{"properties":{"user_id":{"type":"integer","title":"User Id"},"message":{"type":"string","title":"Message"}},"type":"object","required":["user_id","message"],"title":"VoiceRequest"}}}}